Progress on passing arbitrary AD types to the LAMENT library.
The LAME and LAMENT stress evaluators are currently separate classes,
which involves some duplication but simplifies things in the short term.

The LAME stress evaluator uses finite-differencing to compute the Jacobian.

The LAMENT stress evaluator uses Sacado templating to compute the Jacobian.  This
is currently enabled only for the ElasticNew material model.

The LAMENT test is currently disabled because it fails.