MOR: Enabled truncation of projection basis
Added the option to specify the maximum numbers of vectors from a basis
file. Leading vectors (corresponding to higher weights) are
favored and trailing vectors (with lower weights) are ignored.

Refactored basis reading code to eliminate redundancy.

Known defects:
1) Partial basis reading is inefficient. The file basis is first
fully loaded in a MultiVector before a part of it is copied into another
MutiVector. Current public interfaces in the Epetra package impose the
above workflow.
2) The projection basis is duplicated when performing a projection-based
ROM computation. Once copy is used in ReducedOrderModelEvaluator, and
the other in FullStateReconstructor. Each copy is read from disk
separately. This misfeature was already present before the current
changes. It minimizes intrusivity but should be solvable
at little cost.