MOR: Enabled projection-based ROM for Rythmos
The implementation is based on a wrapper for Rythmos::StepperBase which
performs the expansion of the reduced coordinates.
The computed full solution and velocity are then exposed to the wrapped
observer, since the latter expects to be given vectors expressed
in the coordinates of the full-order model.

Providing an implementation for the broad abstract interface of
Rythmos::StepperBase might be overkill but it is necessary to
minimize intrusivity and coupling with Albany internals.
Note that most methods of Albany_RythmosStepperFullStateWrapper
are only stubs and are not meant to be called. Currently, only the ones required in
Albany_RythmosObserver can be expected to be properly implemented.