MOR: Added collocation for mimimum residual
The current implementation is intended to be a reference until a more
efficient alternative is developed.

As in the Galerkin projection case, the entries of the left projector
corresponding to the ignored degrees of freedom are set to zero.
However, the left projector used in the context of the minimum residual
system reduction is not fixed, so that the above operation must be
performed every time the basis changes.