Added a trigonometric source class to PHAL_Source_Def.hpp.
Added a new Steady Heat 2D (Poisson) test case with trigonometric
source.  The source is such that the exact solution to this problem is

u = sin(2*pi*x)*cos(2*pi*y).

The purpose of this test case will be to study the convergence of
finite elements used in Albany.  The test case has not been tested yet.
It will be tested when Irina returns from vacation August 6.