Add some #include guards for ALBANY_FELIX around the transformMesh() method
The transformType is not being initiallized properly (or does not have a
proper default), resulting is undefined behavior (and a ton of debugging).
Essentially, transformMesh() was being called for every LCM problem, with a
random number populating the transformType variable, occasionally leading to a
transformation that would differ from processor to processor in parallel.