Refactor related to multi-block support for LCM problems: impacts all problems
The state field manager created in Albany_Application needs to be aware of
element block info. More specifically, element blocks may have separate state
dependency trees, but current capabilities do not account for this. To address
this issue registration of state variables now includes the name of element
block. This requires a signature change to the registerStateVar method.