Added basic support for non-contiguous node numbering
The change has no effect for contiguous node numberings, or when an
interleaved ordering of the dofs is requested.

With the current implementation of the node-to-dof mapping, a
non-contiguous node numbering implies a non-contiguous numbering for the
global dof indices. Although such an occurrence should be supported by
Epetra, it might nevertheless lead to corner case bugs.