Added timers to Response function calls. Changed timer invocation to use the TEUCHOS_TIME_MONITOR macro.
Now that responses are timed, we can get more info on the
performance issues with large schrodinger-poisson problems
that appear to stall during response calculaions.

The sum of fill times should now be very close to total
fill time.

Slight changes to names of SG timers to conform to the
formatting of the rest of the timer names.