Added a switch within Albany::Application and Albany::ModelEvaluator to turn on/off the computation of sensitivities, regardless of the number of parameters given to the contained Problem.  The switch essentially toggles whether the Output Args of the Problem's model evaluator (Albany::ModelEvaluator) support sensitivities.  The switch is specified at the top level of the Problem's parameter list within the xml input file via the parameter:
<Parameter name="Compute Sensitivities" type="bool" value="1" />

as exemplified by the examples/Poisson/input_mosdot_2D.xml file.  This xml parameter name & location were chosen based on its prior existence as an allowed parameter that was only used in SG problems where it determined whether DgDp_sg was set for the outer ModelEvaluator's outArgs or not (within Main_SGSolve.cpp).  We probably want to consider how to reconcile the two approaches of controlling sensitivities computation in the future.