LCM: revert Lame velocity gradient to used the eig_spd_cos algorithm
Something is not working between the SVD calculation and Lame when calcualting
the velocity gradient. I don't know what it is, but since Lame doesn't need AD,
I am adding another method in the tensor library polar_left_logV_lame that is
specifically to be used in the Lame calculation of the velocity gradient.

Using this we could figure out why the SVD version is failing, but I haven't
done that yet, and I don't have much interest in it either.