Added a new QCAD response: "Field Average"  which evaluates the average value of a field in a given mesh region.  See input_mosdot_2D.xml for example usage.  This has only been tested lightly thus far.
Added a MeshRegion utiltiy class used by several QCAD evaluators (field integral, field value, field average, center of mass, region boundary) to encapsulate the region/area of the mesh the operate on.  This centralizes the code used to determine whether a mesh cell lies in a specified region.  As a result, there are a few slight changes to the above responses parameter lists:
   - field value:  "Operation Domain" is now deprecated and no longer does anything
   - region boundary:  "Region Type" is now no longer a parameter (will result in validation error if used)
A resulting benefit is that now all five of the above responses define the mesh regions they operate on in the same way and have identical capabilities for defining such regions.