Enabled access to history of solution from Exodus
This is a minimally intrusive, quick and dirty method to support reading
arbitrarily large sets of vectors bundled with the input mesh in an
Exodus database.

It suffers from several weaknesses such as lack of generality and
robustness.

For instance, only the field called "solution" can be currently read.

Furthermore, fetching the solution history clobbers the contents of the
fields declared as transient in the Stk database. In particular, the
initial condition becomes unavailable thereafter.