MOR: Enabled support for Stk-based basis input
This change required a deep overhaul of the MOR factory framework to
avoid depending on any Albany specifics. A negative side effect is to
embed the MOR support more deeply into the Albany application.

The main idea is that the necessary resources (e.g. access to the Albany
Stk database) are passed to the MOR components through their
constructors. This technique is an example of dependency injection.