Adding new option to Albany ParameterList, called "Debug Output".  The options under this section are the following at the present time:
 <ParameterList name="Debug Output">
     <Parameter name="Write Jacobian to MatrixMarket" type="int" value="-1"/>
     <Parameter name="Write Residual to MatrixMarket" type="int" value="-1"/>
     <Parameter name="Write Jacobian to Standard Output" type="int" value="1"/>
     <Parameter name="Write Residual to Standard Output" type="int" value="3"/>
  </ParameterList>

The first two will write a selected (value = "N" for N = 1,2,3...) Jacobian and/or
residual to a MatrixMarket file.  The second two will write a selected (value = "N",
for N=1,2,3...) Jacobian and/or residual to the standard output (cout).  A value of
-1 indicates the Jacobian and/or residual is requested each time it is computed in the code.

For an example of "Debug Output"'s usage, please see:
examples/FELIX_Stokes/input_fo_ismip-hom_testB.xml .