MOR: Shared reduced space creation logic
Moved the responsibility of computing the dual projector from the
ModelEvaluator factory up to the ReducedSpace factory.

The goal is to make the corresponding logic consistently available to
the other parts of the framework, in particular the Observer factory.

Known limitation: Computations of the dof sampling and the dual
projector may be performed several times.