MOR: Fixed reduced initial conditions
The reduced components of the initial conditions are now consistently
computed when dof-based sampling is activated.

The projection is defined as
x_r = projector^T * x
such that projector = M * basis * (basis^T * M * basis)^{-1}
where M is the gappy metric corresponding to the sample dofs.

This definition is the same when the full or a reduced mesh is used.