Used Piro solve driver for model evaluation
Made computing the sensitivities optional. The behavior is configured
via a boolean flag labeled "Compute Sensitivities" in the input file,
located under the "Problem" parameter sublist. This choice maintains
backward compatibility, but is likely to be changed in the near future.

Caused minor side effects:
1) Shortened scope for the setup timer
2) Reordered console output

This patch relies on features recently introduced in the Piro package,
and therefore requires an up-to-date Trilinos distribution to compile.