Add install support for the Albany libraries
In suport of an LDRD, we need to be able to locate certain Albany libraries.
I thought it would be good to be able to use the CMAKE find_packge() macro.
This push enables that, however I have only tested it in the context of the
one aforementioned project.

NOTE: This requires CMAKE version > 2.8.8. Only the install portion is now
optionally dependent on this version.

For any further info please contact me (jtostie@sandia.gov)