Deactivated nonterminating examples
On my machine (MacOs X 10.6 / GCC 4.2.1), the tests involving Dakota
restart files run forever. The reason is that one for the 4 MPI
processes raises a std::bad_alloc exception, thus leading to a deadlock
when this single process gets stuck in the catch clause in the main
function, while the other processes still try to follow the normal flow
of control. The bad_alloc is due to a memory allocation request with an
inordinately large size, which I have traced back to a value read from
the user-provided binary file examples/SteadyHeat1D/dak.restart.

This diagnostic can be confirmed by simply running the
dakota_restart_util executable on that file, which also raises a
std::bad_alloc exception.

The reason for this data corruption is still to be determined. A possible
hypothesis is the lack of portability of the binary file format used for
the restart file.

Until the root cause is found and fixed, I will leave the corresponding
tests deactivated.