Refactored regression testing support
1) Broke up monolithic testing function into four specialized functions
that handle different types of runs (Solve, Analysis, Dakota and SG).

2) Enabled detection of missing quantities in tests.

A regression test requested for a missing quantity (denoted by a null
pointer) used to be simply skipped. However, such an approach did not
allow to detect whether any user-requested quantity was not computed.

The new implementation automatically fails a test with the same error
code as the one corresponding to a size mismatch when a quantity
requested to be tested is absent.