Changed implementation of Compute Sensitivity flag
The Albany ModelEvaluator used to inhibit the support for the various
derivatives objects whenever the "Compute Sensitivity" boolean parameter
was explicitly set to false.
The driver functions then automatically computed the sensitivies if and
only if the model they were given was prentending to support them.

Support for the derivatives in the Albany ModelEvaluator does not depend
on the value of the flag anymore: It is now always enabled.
Additionally, the flag has been repurposed to directly select the
behavior of the driver functions.