Decomposed observer factory
Publicly exposed individual object functions encapsulating the creation
logic for each type of Observers.

The monolithic factory is now implemented in terms of the above
functors. It also includes some logic that determines what kind of
observer should be constructed depending on which type Piro Solver will
be created by Albany. This logic is redundant with the one present in
SolverFactory and is a prime candidate for refactoring.