Delegated Solver creation to Piro
Removed the Albany-specific logic dedicated to the creation of the Piro
Solvers.

Instead, Piro::Epetra::SolverFactory is used to drive the whole process.
The delicate problem of creating the Albany-specific auxiliary objects
(e.g. Observers) to be passed as arguments to the constructors of the
Piro solvers is handled by extending the Piro Solver Factory with the
appropriate callbacks (called 'Providers').

The main benefit is that Albany is now independent from the exact types
and the implemention details of the different Piro Solvers.
The structure of the Piro parameter list is the only form of coupling
remaining.

Note: This changeset requires to update Trilinos to obtain a recent
version of the Piro package.