LCM and Adapt: changes stemming from code review to Topology class
Alejandro and I took a first pass through the LCM::Topology class after a
a recent coding session with Glen. Some minor functionality was changed, but
most of the changes are cosmetic to increase readilbility and consistency with
the style guidelines the LCM team is trying to follow (for the record, we have
adopted the google c/c++ style).