Enabled Thyra-based NOXSolver as a runtime option
Added an example based on SteadyHeat1D that demonstrates the new
feature.

The Thyra-based Piro::NOXSolver can be used as an alternative to its
Epetra equivalent.

The feature is available as a runtime option by setting the "Solver
Type" string parameter located directly under the Piro section of the XML
input file to the value "ThyraNOX". In that case, the "NOX" parameter
sublist is used to setup the Thyra solver.

Currently the NOXSolver can only be used in the Solve driver. Making it
available in other parts of Albany would likely require to create an
subclass of EpetraExt::ModelEvaluator that would wrap a Thyra model
evaluator.