LCM: Initial MaterialPointSimulator test
At this point the MaterialPointSimulator (MPS) can be used to test constitutive
models in the src/LCM/models directory through the ConstitutiveModelParameters
and ConstitutiveModelInterface evaluators. This commit adds the first test, and
it exercises the AnisotropicHyperelasticDamage model in uniaxial tension. There
are a few things to note:

1) The MPS writes an exodus file for the stress and state variables of the model
2) To check the test, we are invoking exodiff to check all components of all
quantities at all steps against a gold file
3) Both the MPS executable and exodiff are being symlinked to the test directory
4) a python script is being used to run the MPS and exodiff, and ctest is just
calling the python script.

more to come.