MOR: Added regression test for the hyperreduction
The test checks that the value of the solution at the sensor node, which
is an arbitrary chosen sample node, matches the expected value given by
the truth, i.e. the Galerkin-projection-based ROM without Missing Point
Estimation (MPE) hyperreduction.

Note the input mesh file used here is the same as the truth, updated
with the relevant nodeset information. Since there is no mesh reduction,
no efficiency gain is expected.

The example also provides a regression test for a newly introduced type
of response which selects the dofs of the solution vector located at the
nodes belonging to a user-designated nodeset.