Fixed conversion of locally replicated maps
For parallel runs (i.e. when the number of processes is strictly greater
than one), locally replicated maps correspond to a special case that
requires a special treatment.

In that case, Epetra maps must be created as instances of the subclass
Epetra_LocalMap.
Likewise, Tpetra maps must be constructed with a particular flag value
passed as an input argument.

Also, made the conversion more robust:
1) Checked for overflow when converting between different index types
2) Enabled non-zero base index