LCM: Add BifurcationCheck Evaluator to check material stability
This is just a shell that I am providing to Alejandro and Qiushi. Currently,
the MaterialPointSimulator will parse it's input for a 'Check Stability' flag,
and if that is true it will pass the material tangent to the BifurcationCheck
Evaluator, which returns if the point is elliptic, and a direction.