Improved configuration logic for MueLu examples
The MueLu examples require that the Trilinos installation include two
optional dependencies of this package, namely Ifpack2 and Amesos2.

Furthermore, the default solver for Amesos2 (Klu) is currently not
available, hence a fallback option, namely the third-party library
SuperLU, must be enabled as well.

If MueLu is available but any of the above conditions is not fulfilled,
then the MueLu examples will compile properly but will fail at runtime.

The configuration script tries to detect such an occurence and enables
the MueLu tests only if they can be expected to succeed.