Revised configuration of nightly test harness:
*) Added missing dependencies:
   - Zoltan is required to build Isorropia;
   - Some examples rely on SEACASEpu and SEACASExodiff.

*) Activated Ifpack2 to enable corresponding Tpetra/Thyra examples.

*) Removed extraneous dependencies (e.g. RBGen and Hdf5).

*) Cleaned up and reorganized configuration scripts; Improved
consistency.

*) Deactivated MOR in Albany: Some regression tests are based on a
response that has not been converted to Tpetra yet and therefore the
corresponding examples are expected to fail.

Note 1: The serial build is currently broken since several core and optional
components of Albany (directly or indirectly) depend on MPI.

Note 2: MueLu support in Albany is properly configured and compiled, but
the corresponding examples are not run. They would additionally require
Amesos2 with the SuperLU adapter enabled. However, that option is not
activated since the setup complexity outweighs the benefits. For the
time being, Ifpack2 is the preferred preconditioner for Tpetra.