MOR: Added driver for reduced basis generation
The driver is built only when the Trilinos pacakge RBGen is available
and the MOR support in Albany has been enabled.

The driver duplicates a chunk of logic from Albany::Application that
creates and initializes a problem instance. While such an approach is
likely to be overkill, making the implementation leaner would probably
require significant refactoring.