MOR: Added driver for mesh sampling and reduction
The driver is built only when both the MOR and SECAS/Exodus support in
Albany has been enabled.

The driver duplicates a chunk of logic from Albany::Application that
creates and initializes a problem instance. While such an approach is
likely to be overkill, making the implementation leaner would probably
require significant refactoring.