Introduced stub Tpetra-based AdaptiveSolutionManager
The goal is to start bridging the differences between the master and the
Tpetra branches.

Since there is no Tpetra-based or Thyra-based counterpart for the
Epetra AdaptiveSolutionManager hierarchy in packages NOX and Piro, the
newly introduced class cannot derive from those and therefore offers no
adaptation capabilities. It only provides basic functionalities, such as
handling the conversion between non-overlapping and overlapping vectors.