Factored common global residual computation
The two public methods (Epetra-based and Tpetra-based) computing the
global residual used to have separate implementations, with significant
duplication.

After refactoring, both call a common private Tpetra-based method.

The implementation of the Tpetra-based setResidualT method is still
missing in STKDiscretization. A temporary workaround is to perform a
conversion before calling the Epetra version.