Converted TransientHeat2D example to Tpetra
This test case demonstrates the Thyra-based Piro::RythmosSolver on a
more challenging problem than the simple ODE.
In particular, it is a distributed problem that requires the computation
of the sensitivities wrt the parameter.
Additionally, it provides another use case for the Belos linear solver
combined with the Ifpack2 preconditioner.