MOR: Set default basis size to snapshot count
RBGen uses an arbitrary value (namely, 16) when the target number of
basis vectors is not explicitly specified. That behavior is
counterintuitive and should be avoided.

Therefore, whenvever the target number of basis vectors has not been
specified in the input file, the corresponding entry in the parameter
list is explicitly set equal to the (modified) snapshot count before it
is passed to RBGen.