MOR: Enabled dropping leading vectors in mesh sampling
The goal is to be able to ignore the first vector read from the input
file history via the stk mesh database when it is not part of the
orthogonal reduced basis but rather corresponds to the origin and
therefore should not be used to compute the sample mesh.

This is mostly a temporary solution. Ultimately, the same logic (and
input file syntax) that creates the reduced space elements should be
used here instead.