Converted SurfaceElement example to Tpetra
Extended time table in the Tpetra version to avoid out-of-range lookups.

The latter seem to occur during the computation of a derivative with
respect to time (which is the continuation parameter in this particular
case) via a finite difference scheme. I am not sure why such a problem
happens only in the Tpetra version.