Converted ThermoMechanical3D example to Tpetra
Switched to natural continuation method to ensure that the problem is
solved via pseudo-time stepping with constant steps.
For some reason to be determined, the Thyra version of the Arc Length
method does not behave like the Epetra implementation and takes an extra
step with a smaller increment, leading to a slightly but significantly
different final solution.

Extended time table to avoid out-of-bounds lookups.

The resulting Exodus output files do not exactly match the ones produced
by Epetra runs. The residual for the displacement in the z direction
exhibits the only numerically significant discrepancy. All other results
do match within a 1.0e-9 relative tolerance.