Converted Mechanics2D examples to Tpetra
For the continuation problem involving J2 plasticity model, while the
regression test succeeds, the Exodus output file obtained with the
Tpetra version differs significantly from one produced using the Epetra
reference.

The most striking discrepancy relates to the stamps of the transient
states. The Tpetra version does not seem to correctly output the
continuation parameter value, instead using the rank of the step as a
default value.

The only nodal variable for which the relative discrepancy is not
negligible is the residual of the displacement in the y direction.
However, the magnitude is small enough compared to the x component to be
ignored.

On the other hand, some element variables exhibit significant
differences. In particular, the value of the 4 components of
Mechanical_Source is zero for the Tpetra version, but not for the Epetra
reference, for which all components share the same non-null value, about
16.2. Furthermore, most entries of Cauchy_Stress and F fields show a
significant relative discrepancy, but the magnitude of these are small
compared to the few entries that match (namely 1, 5, 9 and 13 for
Cauchy_Stress, and 1, 4, 5, 8, 9, 12, 13 and 16 for F).