Removed obsolete conditional compilation of output
It should always be legal to call the pure virtual member function
AbstractDiscretization::writeSolution independently of the compile-time
configuration and the available packages.

The concrete implementations of AbstractDiscretization (including the
STK one) are now responsible for detecting the absence of I/O support
and protect against compilation and run-time errors.