Broke cycle of strong references
There was a cycle a strong references involving the Piro solver,
Albany::ModelEvaluator, Albany::Application, AdaptiveSolutionManager,
AdaptiveModelFactory, and the Piro solver again, closing the loop.

One possible solution is to break the above cycle by replacing one
strong (owning) RCP by a weak (non-owning) one. Such a change is always
risky since it can introduce dangling references, and therefore requires
to make an hypothesis concerning the relative lifetimes of the involved
objects.

Here is is assumed that the adaptivity component will not try to
access the Piro solver after the latter has been destroyed.
Bad things may happen if this assumption is violated.