Extended Ascii mesh reader to work in parallel.
For a parallel run, 7 Ascii files are required on each processor N:

1.) xyzN: file giving x,y,z coordinates of mesh
2.) elesN: file giving finite element connectivity
3.) bfN: file giving basal face connectivity (for FELIX)
4.) shN: file giving surface height values at each node (for FELIX)
5.) gnIDsN: file giving global node #s to all the nodes in xyzN file
6.) geIDsN: file giving global element #s to all the elements in the mesh
7.) bfIDsN: file giving global face #s to all the basal faces on the basal boundary.

Each file has a header as a the first line that gives the total # rows in the file (not including the header).

bfN and shN are only needed for FELIX problems.

All global IDs are 1-based, not 0-based.

Usage for Ascii mesh reader should remain the same as before for serial runs.  In particular,
files 1.)-4.) are names xyz, eles, bf, and sh for a serial run, and files 5.)-7.) are not required.

To understand better the usage of the Ascii mesh reader in parallel please see the following new example:

examples/FELIX_Stokes/AsciiMeshes/Gis20km