MOR: Switched to 2-norm sample selection criterion
Replaced hardcoded sample selection criterion based on the Frobenius
norm to one based on the 2-norm (equivalently, spectral radius or
largest eigenvalue).

The new criterion is slightly more involved and possibly more expensive
to compute but gives better results since it improves the "worst"
direction, which is characterized by the eigenvalue of the discrepancy
matrix with the largest magnitude.