Added the capability to request mesh coordinates to be written to a Matrix Market file from the .xml input file.  To request the mesh coordinates to be written to a Matrix Market file, the following line should be added under the "Discretization" section of the input file:
 <Parameter name="Write Coordinates to MatrixMarket" type="bool" value="1"/>

NOTE: ML must be selected as a preconditioner for this line to work.  This is because the coordinates
are primarily needed in ML analysis, and this was a feature request for ML analysis.

For an example input file that calls the new feature, see:

examples/FELIX_Stokes/FO_ISMIP/input_fo_ismip-hom_testB.xml