Added some Performance Tests: first attempt
Thanks to Dave Littlewood who created this capability, in Peridym.

There is a new directory:  examples/PerformanceTests
that has a directory for each test. In that directory is
a "data.perf" file that has "gold standard" time that is
labeled by machine name, along with an acceptible tolerance
on that time, e.g.

upenn           1                     1.70                0.05
redsky          1                     2.70                0.15

Feel free to add your machine to the data.perf files.
The machine name is the result of
 % hostname -s
and the max_wallclock_time is from this output line in Albany:
  Albany: ***Total Time***               1.702 (1)

Currently the executable is hardwired to be Albany on 1 processor.