MOR: Avoided copy in MV truncation
Multivector truncation used to be implemented by allocating a smaller
multivector then making a copy of the data of the leading vectors.

Truncation is now achieved by creating a view of the leading vectors.
Lifetimes are managed using advanced RCP machinery.

The new approach is not memory efficient. The former behavior can be
recovered by making a copy of the returned view before letting the
latter be reclaimed.