Adding FELIX 20km Greenland example where the Albany Adaptivity class is used to do a uniform mesh refinement of an input exodus mesh (gis20km_upn4_in.exo).  Unlike the gis20km_in.exo mesh, this mesh has only 4 levels in the vertical (z-) direction.  The level of refinement can be changed in the input file input_fo_gis20km_refine.xml by modifying the line
 <Parameter name="Number of Refinement Passes" type="int" value="1"/>

to indicate how many levels of refinement are desired.

Note that for the uniform mesh refinement to work, Albany needs to be build with ALBANY_STK_PERCEPT
enabled:

-D ALBANY_STK_PERCEPT=ON

For this to work, Trilinos needs to be build w/ some extra packages enabled, in particular
with a version of Boost built with extra libraries.  See:

http://redmine.scorec.rpi.edu/projects/albany-rpi/wiki/Wiki