LCM: Worked around linking bug in STK package
The static member constants NODE_RANK, EDGE_RANK and FACE_RANK are not
properly defined outside of the enclosing class and therefore lack
external linkage. In other words, they are *not* defined in the
resulting object file.

On my machine, the problem only occurs in debug mode; I suspect that in
release mode the above constants are optimized out and simply replaced
by their values.

Until the proper declarations are introduced, any reference to such
constants should be avoided, and the corresponding member functions
should be used instead. These functions are likely candidates for inline
expansion in release mode so runtime performance should not be affected.