LCM: fix issues, bugs, mistakes in the Mechanics problem with temperature
Where to start...The nice people over at RPI wanted to do a thermo-mechanical
film on a substrate type problem, and tried to use capabilities in the LCM
Mechanics Problem. Turns out that wasn't doing what it was supposed to, and
digging uncovered some serious errors in the design strategy that I (clearly)
didn't finish. So I have some humble gratitude to them for exposing those
issues.

To fix this, the Constitutive Model Interface is now aware of an optional
temperature field, and the Constitutive Model class deals with storing some
additional material properties. It is still up to the Model implementation
to provide the temperature dependent stress, for example the adjustment due
to thermal expansion.

So I added a test with no applied loading, just a temperature ramp, and the
Neohookean model ccomputes a corresponding volume change.

There is more testing to be done for this...