Towards shallow water on the sphere.
Steve Bova has implemented the shallow water
equations on a 2D plane surface. This is to
be extended to the shallow water equations
on a sphere. The sphere is 3D but the parametric
or reference element is a 2D quad. So far all
of the changes are based on the dimension of
the problem, 2D or 3D in order not to break
the plane solution.

This commit sets up the sphere problem example
but fails in the solver routine.  Likely because
the 2D/3D mix in a spherical quad element is
causing problems.