Changed CismSTKMeshStruct constructor so it takes in arrays rather than std::vectors, so that arrays passed from Cism do not need to be copied into std::vectors in felix_driver.cpp .
Added code to copy solution computed in Albany into uvel and vvel arrays to be passed
back to CISM.  This isn't quite right yet -- need to figure out size / structure of
uvel and vvel arrays in CISM and make sure solution is copied according to the CISM order
of the solution dofs.

More cleanups of the interface code.