Added x_dotdot support in Albany, for proper treatment of second derivatives in time.
Note: These changes are compatible with changes in Trilinos, so
both must be updated at the same time.

Transient Mechanics applications need time integrators for x_dotdot.
An earlier hack to allow this was no longer sufficient.

Unfortunately, many of the changes here are associated with
responses, directional derivatives, stochastic Galerkin, or
multi-point evaluations, and these are not tested.