corrected comment changed Sphere10 example to go about a quarter way around the sphere.
Sphere10 example now advects a bump around the sphere.  This works
with both linear and quadratic polynomials on quads.

Added handling for quadratic shealls on the sphere.

THere was some confusion in the code about coordinates living at
vertices versus nodes.  I removed the vertices_3vector and replaced it
with a node_3vector type.  I suspect there may be other bugs elsewhere
in Albany where the coordinates are using the vertices_vector instead
of the nodes_vector layout.