LCM: CHANGING DEFAULT OUTPUT BEHAVIOR
There were many places in the code where state variables were being registered,
whether or not they needed to be, generating a lot of extra memory usage and
generally cluttering up the output files. Now, by default, much of those
integration point quantities are not registered as state variables, and in
fact are only registered as state variables if requested for output in the
input file. There is more work to do here. The input file syntax looks like
the following:

<Parameter name="Output Field" type="bool" value="true/false"/>

where Field in the name of the integration point field.