LCM: finish generalizatio of pusing ip quantities to nodes
Currently, there is an issue with trying to specify one response per field.
Not sure why but that doesn't work as expected. Possibly a better way to
compute multuple nodal fields is by specifying multiple integration point
fields in the response block. This has the added benefit of only having to
compute the nodal weights once. The syntax has been updated in the KfieldBC
example, and is summarized below for the J2 model and outputting stress and
eqps.

      <ParameterList name="ResponseParams 2">
        <Parameter name="Number of Fields" type="int" value="2"/>
        <Parameter name="IP Field Name 0" type="string" value="Cauchy_Stress"/>
        <Parameter name="IP Field Layout 0" type="string" value="Tensor"/>
        <Parameter name="IP Field Name 1" type="string" value="eqps"/>
        <Parameter name="IP Field Layout 1" type="string" value="Scalar"/>
        <Parameter name="Output to File" type="bool" value="true"/>
      </ParameterList>