Beginning implementation of test case 5 in Aeras + some code cleanups:
- Renamed "TestCase2" initial condition "Aeras ZonalFlow".  This is because this same condition
will be used also in test case 5.  Added one additional argument to the specification of the data array
for this boundary condition which is the value of h0*g.

- Fixed Aeras tests so they all pass when you run ctest from the Aeras directory.  This required
putting a CMakeLists.txt in the grids directory so the *.g file the test cases rely on gets
put in the build tree, and putting the correct path to this file in the *.xml input files.

- Created input file for TestCase5.

- Wrote a Surface Height evaluator for the case when h* \neq h (h* = h - hs), as is the case for
test case 5.  This does not work yet correctly, as the coordVec layout is not correct.  To fix this I need
to discuss a few things with the group about the transformation of coordVec onto the spherical meshes/its data layout.
Test cases for which hs = 0 (h* = h) should be unaffected by the new surface height evaluator and its incomplete state.