Adding K-field example with localization  elements and hydrogen diffusion
This is a first step towards mking this problem more robust

What is needed:

- Initial conditions need to be established by element block
- ML preconditioner or changes to the linear solver that speed things up

The initial conditions will require a little code and changing all the tests. I need to do that. With regard to ML and the linear solver -> this problem works ok but finer meshes do not scale well. Finer meshes are required.